Restructures utils folder
