+++ src/components/Box.jsx
@@ -1,5 +1,5 @@
 import styled from 'styled-components'
-import applyStyles from '../utils/applyStyles'
+import applyStyles from '../utils/styles/applyStyles'
 
 const Box = styled.div`
   ${(props) => applyStyles(props)};
