+++ src/utils/breakpoints/openBreakpoint/index.js
@@ -1,0 +1,2 @@
+export { default } from './openBreakpoint'
+export * from './openBreakpoint'
