+++ src/utils/functions/pick/pick.spec.js
