+++ src/utils/functions/pop/pop.js
