+++ src/utils/functions/pop/pop.spec.js
