--- src/utils/math/toNumber.js
@@ -1,4 +1,0 @@
-// @flow
-export default function toNumber(value: any): number {
-  return Number.isInteger(value) ? value : parseFloat(value)
-}
