+++ src/utils/strings/getPrefix/getPrefix.js
