+++ src/utils/strings/getPrefix/index.js
@@ -1,0 +1,2 @@
+export { default } from './getPrefix'
+export * from './getPrefix'
