+++ src/utils/styles/createMediaQuery/createMediaQuery.js
@@ -2,9 +2,9 @@
 import type {
   TBreakpoint,
   TBreakpointBehavior,
-} from '../../const/defaultOptions'
-import transformNumeric from '../math/transformNumeric'
-import toDashedString from '../strings/toDashedString'
+} from '../../../const/defaultOptions'
+import transformNumeric from '../../math/transformNumeric'
+import toDashedString from '../../strings/toDashedString'
 
 const shouldAppendProp = (propName: string, behavior: TBreakpointBehavior) => {
   const [prefix, splitPropName] = propName.split('-')
