+++ src/utils/templates/generateComponents/index.js
