+++ src/utils/templates/getAreasList/index.js
