+++ src/utils/templates/parseTemplates/filterTemplateProps.spec.js
